extrator -> extractor
