+++ pic_tag/camera_worker/feature_extractor/ReIDAttenv2_14999.pth
