+++ pic_tag/camera_worker/feature_extractor/ReIDAttenv2_aug_6000.pth
