+++ pic_tag/camera_worker/feature_extractor/ReID_model.py
