+++ pic_tag/camera_worker/feature_extractor/feature_extractorv2.py
